Make Xwayland process more elegant
This also:
- Handle Xwayland startup failures
- Makes the auto start slighly more reliable

+++ src/server/frontend_xwayland/xwayland_server.cpp
@@ -55,18 +55,21 @@ mf::XWaylandServer::XWaylandServer(const int xdisplay, std::shared_ptr<mf::Wayla
 
 mf::XWaylandServer::~XWaylandServer()
 {
-    mir::log_info("deiniting xwayland server");
+    mir::log_info("Deiniting xwayland server");
 
     // Terminate any running xservers
-    kill(pid, SIGTERM);
+    if (xserver_status > 0) {
+      terminate = true;
 
-    if (lazy) {
-      lazy_thread->join();
-      return;
+      if (xserver_status == RUNNING)
+        wm->destroy();
+
+      kill(pid, SIGTERM);
+      spawn_thread->join();
     }
 
-    dispatcher->remove_watch(afd_dispatcher);
-    dispatcher->remove_watch(fd_dispatcher);
+    if (lazy)
+      return;
 
     char path[256];
     snprintf(path, sizeof path, "/tmp/.X%d-lock", xdisplay);
@@ -81,7 +84,16 @@ void mf::XWaylandServer::spawn()
 {
     int fd;
     int wl_client_fd[2], wm_fd[2];
+    int status;
     std::string fd_str, abs_fd_str, wm_fd_str;
+
+    // Try 5 times
+    if (xserver_spawn_tries >= 5) {
+      mir::log_error("Xwayland failed to start 5 times in a row, disabling Xserver");
+      return;
+    }
+    xserver_status = STARTING;
+    xserver_spawn_tries++;
 
     if (socketpair(AF_UNIX, SOCK_STREAM | SOCK_CLOEXEC, 0, wl_client_fd) < 0)
     {
@@ -101,11 +113,11 @@ void mf::XWaylandServer::spawn()
 
     // This is not pretty! but SIGUSR1 is the only way we know the xserver
     // is ready to accept connections to the wm fd
-    // If not it will have a rase condidion on start that might end
+    // If not it will have a race condition on start that might end
     // up in a never ending wait
     std::signal(SIGUSR1, [](int) {
         xserver_ready = true;
-        mir::log_info("stated");
+        mir::log_info("Xwayland running");
     });
 
     mir::log_info("Starting Xwayland");
@@ -136,36 +148,86 @@ void mf::XWaylandServer::spawn()
 
         // forward SIGUSR1 to parent thread (us)
         signal(SIGUSR1, SIG_IGN);
+
+        // Last second abort
+        if (terminate) return;
+
         if (lazy)
-            execl("/usr/bin/Xwayland", "Xwayland", dsp_str.c_str(), "-rootless", "-wm", wm_fd_str.c_str(), NULL);
+            execl("/usr/bin/Xwayland",
+                  "Xwayland",
+                  dsp_str.c_str(),
+                  "-rootless",
+                  "-wm", wm_fd_str.c_str(),
+                  NULL);
         else
-            execl("/usr/bin/Xwayland", "Xwayland", dsp_str.c_str(), "-rootless", "-terminate", "-listen",
-                  abs_fd_str.c_str(), "-listen", fd_str.c_str(), "-wm", wm_fd_str.c_str(), NULL);
+            execl("/usr/bin/Xwayland",
+                "Xwayland",
+                dsp_str.c_str(),
+                "-rootless",
+                "-listen", abs_fd_str.c_str(),
+                "-listen", fd_str.c_str(),
+                "-wm", wm_fd_str.c_str(),
+                "-terminate",
+                NULL);
         break;
     case -1:
         mir::log_error("Failed to fork");
         break;
     default:
         close(wl_client_fd[1]);
+        close(wm_fd[1]);
         auto wlclient = wl_client_create(wlc->get_wl_display(), wl_client_fd[0]);
 
-        close(wm_fd[1]);
-
         // More ugliness
-        // TODO handle xserver startup failures!
-        mir::log_info(xserver_ready ? "xs true" : "xs false");
+        int tries = 0;
         while (!xserver_ready)
         {
-            std::this_thread::sleep_for(std::chrono::milliseconds(10));
+          // Last second abort
+          if (terminate) return;
+
+          // Check for stalled startup
+          if (waitpid(pid, NULL, WNOHANG) != 0 || tries > 200) {
+            xserver_status = FAILED;
+            mir::log_info("Stalled start of Xserver, trying to start again!");
+            new_spawn_thread();
+            return;
+          }
+
+          std::this_thread::sleep_for(std::chrono::milliseconds(10));
+          tries++;
         }
-        mir::log_info(xserver_ready ? "xs true" : "xs false");
 
-        int status;
-
+        // Last second abort
+        if (terminate) return;
         wm->start(wlclient, wm_fd[0]);
+        mir::log_info("XServer is running");
+        xserver_status = RUNNING;
+
+        // Reset the tries since server is running now
+        xserver_spawn_tries = 0;
+
         waitpid(pid, &status, 0);  // Blocking
-        wm.reset();
+        if (WIFEXITED(status)) {
+           mir::log_info("Xserver stopped");
+           xserver_status = STOPPED;
+       } else {
+          // Failed, crash or killed
+          mir::log_info("Xserver crashed or got killed");
+          xserver_status = FAILED;
+       }
+
+        if (terminate) return;
+        wm->destroy();
         xserver_ready = false;
+
+        if (xserver_status == FAILED) {
+          mir::log_info("Trying to start Xwayland again!");
+          new_spawn_thread();
+          return;
+        }
+
+        spawn_xserver_on_event_loop();
+
         mir::log_info("Xwayland stopped");
         break;
     }
@@ -185,14 +247,14 @@ void mf::XWaylandServer::bind_to_socket()
 
     if (bind(socket_fd, (struct sockaddr *)&addr, size) < 0)
     {
-        mir::log_error("Failed to bind to x11 socket");
+        mir::fatal_error("Failed to bind to x11 socket");
         close(socket_fd);
         return;
     }
 
     if (listen(socket_fd, 1) < 0)
     {
-        mir::log_error("Failed to listen to x11 socket");
+        mir::fatal_error("Failed to listen to x11 socket");
         unlink(addr.sun_path);
         close(socket_fd);
         return;
@@ -211,14 +273,14 @@ void mf::XWaylandServer::bind_to_abstract_socket()
     size = offsetof(struct sockaddr_un, sun_path) + name_size;
     if (bind(abstract_socket_fd, (struct sockaddr *)&addr, size) < 0)
     {
-        mir::log_error("Failed to bind to abstract socket");
+        mir::fatal_error("Failed to bind to X11 abstract socket");
         close(abstract_socket_fd);
         return;
     }
 
     if (listen(abstract_socket_fd, 1) < 0)
     {
-        mir::log_error("Failed to listen to abstract socket");
+        mir::fatal_error("Failed to listen to X11 abstract socket");
         close(abstract_socket_fd);
         return;
     }
@@ -241,7 +303,6 @@ int mf::XWaylandServer::create_lockfile()
         {
             if (fd >= 0)
                 close(fd);
-            mir::log_error("Lockfile exists!!");
             return EEXIST;
         }
 
@@ -255,30 +316,20 @@ int mf::XWaylandServer::create_lockfile()
             /* stale lock file; unlink and try again */
             close(fd);
             if (unlink(lockfile))
-            {
-                mir::log_error("Lockfile exists!!");
                 return EEXIST;
-            }
             else
-            {
-                mir::log_error("again!!");
                 return EAGAIN;
-            }
         }
 
         close(fd);
         return EEXIST;
     }
     else if (fd < 0)
-    {
-        mir::log_error("Lockfile exists1!!");
         return EEXIST;
-    }
 
     size = dprintf(fd, "%10d\n", getpid());
     if (size != 11)
     {
-        mir::log_error("Lockfile exists2!!");
         unlink(lockfile);
         close(fd);
         return EEXIST;
@@ -291,23 +342,53 @@ int mf::XWaylandServer::create_lockfile()
 
 void mf::XWaylandServer::setup_socket()
 {
+    char path[256];
     int i = create_lockfile();
     while (i == EAGAIN)
     {
         i = create_lockfile();
     }
-    if (i != 0)
-        mir::log_info("ERROR");
+    if (i != 0) {
+      mir::fatal_error("X11 lockfile already exists!");
+      return;
+    }
+
+    // Make sure we remove the Xserver socket before creating new ones
+    snprintf(path, sizeof path, "/tmp/.X11-unix/X%d", xdisplay);
+    unlink(path);
+
     bind_to_abstract_socket();
+    if (errno == EADDRINUSE) {
+        mir::fatal_error("X11 socket is already in use");
+        snprintf(path, sizeof path, "/tmp/.X%d-lock", xdisplay);
+        unlink(path);
+        return;
+    }
+
     bind_to_socket();
+}
+
+void mf::XWaylandServer::new_spawn_thread() {
+  // Maks sure we don't run the server more then once!
+  if (xserver_status > 0) return;
+  xserver_status = STARTING;
+
+  spawn_thread = std::make_unique<std::thread>(&mf::XWaylandServer::spawn, this);
 }
 
 void mf::XWaylandServer::spawn_xserver_on_event_loop()
 {
-    afd_dispatcher =
-        std::make_shared<md::ReadableFd>(mir::Fd{mir::IntOwnedFd{abstract_socket_fd}}, [this]() { spawn(); });
+    auto func = [this]() {
+      dispatcher->remove_watch(afd_dispatcher);
+      dispatcher->remove_watch(fd_dispatcher);
+      afd_dispatcher.reset();
+      fd_dispatcher.reset();
 
-    fd_dispatcher = std::make_shared<md::ReadableFd>(mir::Fd{mir::IntOwnedFd{socket_fd}}, [this]() { spawn(); });
+      new_spawn_thread();
+    };
+
+    afd_dispatcher = std::make_shared<md::ReadableFd>(mir::Fd{mir::IntOwnedFd{abstract_socket_fd}}, func);
+    fd_dispatcher = std::make_shared<md::ReadableFd>(mir::Fd{mir::IntOwnedFd{socket_fd}}, func);
     dispatcher->add_watch(afd_dispatcher);
     dispatcher->add_watch(fd_dispatcher);
 }
@@ -315,5 +396,5 @@ void mf::XWaylandServer::spawn_xserver_on_event_loop()
 void mf::XWaylandServer::spawn_lazy_xserver()
 {
     lazy = true;
-    lazy_thread = std::make_unique<std::thread>(&mf::XWaylandServer::spawn, this);
+    new_spawn_thread();
 }
